Rename: "view.js -> renderer.js"
해당 파일은 renderer 객체만을 export하기때문에 더 명확한 의미전달을 위해 변경
